feat: introduce agent-control health-check configuration
