+++ agent-control/src/agent_control.rs
@@ -1,16 +1,20 @@
+pub mod agent_id;
 pub mod config;
 pub mod config_storer;
+pub mod config_validator;
 pub mod defaults;
 pub mod error;
-pub(super) mod event_handler;
-pub use agent_control::*;
-#[allow(clippy::module_inception)]
-mod agent_control;
-pub mod agent_id;
-pub mod config_validator;
 pub mod http_server;
 pub mod pid_cache;
 pub mod resource_cleaner;
 pub mod run;
 pub mod uptime_report;
 pub mod version_updater;
+
+pub(super) mod event_handler;
+
+mod health_checker;
+
+pub use agent_control::*;
+#[allow(clippy::module_inception)]
+mod agent_control;
