+++ agent-control/src/agent_type/runtime_config/k8s.rs
@@ -1,7 +1,6 @@
-use crate::agent_type::definition::Variables;
 use crate::agent_type::error::AgentTypeError;
-use crate::agent_type::runtime_config::HealthCheckInterval;
 use crate::agent_type::templates::Templateable;
+use crate::{agent_type::definition::Variables, health::health_checker::HealthCheckInterval};
 use serde::Deserialize;
 use std::collections::{BTreeMap, HashMap};
