+++ agent-control/src/health.rs
@@ -1,7 +1,4 @@
 pub mod health_checker;
-
 pub mod k8s;
-
 pub mod on_host;
-
 pub mod with_start_time;
