+++ agent-control/src/sub_agent.rs
@@ -1,7 +1,12 @@
 pub mod collection;
 pub mod effective_agents_assembler;
 pub mod error;
+pub mod identity;
 pub mod remote_config_parser;
+
+pub(crate) mod event_handler;
+
+mod health_checker;
 
 pub mod k8s;
 
@@ -10,7 +15,5 @@ pub mod supervisor;
 pub mod version;
 
 pub use sub_agent::*;
-pub(crate) mod event_handler;
-pub mod identity;
 #[allow(clippy::module_inception)]
 mod sub_agent;
